requirements.txt
pyautogui
sounddevice
scipy
openai-whisper
pyttsx3
torch
scikit-learn
pandas
numpy
streamlit